+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/entity/video/RelatedVideo.kt
@@ -38,8 +38,8 @@ data class RelatedVideo(
                 author = relate.owner.let { Author.fromVideoOwner(it) },
                 jumpToSeason = false,
                 epid = null,
-                view = relate.stat.view,
-                danmaku = relate.stat.danmaku
+                view = relate.stat.view.toInt(),
+                danmaku = relate.stat.danmaku.toInt()
             )
     }
 }
